requirementstmp.txt
Django==2.2
djangorestframework==3.13.1
pytz==2022.1
sqlparse==0.4.2
